[fxbox/thinkerbell] Moving compilation out of the main thread.
Having compilation on the main thread makes things much more complex,
as we need to move entire graphs off main thread for execution. With
this refactoring, the source ast (which is a much simpler graph) is
moved off the main thread prior to compilation, and we only manipulate
the compiled graph on the dedicated thread.

To get this to work, the API had to be reworked a bit, as some errors
can now be sent asynchronously only.