[fxbox/taxonomy] Resolves #5 - Reworking {Node, Input, Output}Request to make them future-proof.
In the original version, requests were implemented as a `struct` with
`pub` fields. That's going to cause code breakage if we decide to add
or rework fields in the future (e.g. to restore subnodes). This patch
replaces the public fields with a proper API that may be used to
create and combine requests.