[fxbox/taxonomy] Issue #14 (de)serialization of OpenClosed, OnOff
Prior to this patch, empty variants such as "OpenClosed::On" were
serialized as "On: []", which is quite surprising. This patch
overrides the default behavior of OpenClosed and OnOff to make sure
that the empty variants are serialized simply as strings.

For this purpose, the patch introduces a structure TrivialEnumVisitor,
designed to let us easily write parsers for trivial enums.