Removing the big Mutex in taxonomy_router
To be compatible with Iron's thread pool, the taxonomy_router needs to
be Sync. This used to be implemented by putting everything in a big
Mutex.

Changes in Taxonomy have made the AdapterManager Sync and
thread-friendly. This patch propagates the changes to all Adapters
currently in FoxBox and to the taxonomy_router itself.