Resolves #502 - Send/Fetch now require selectors that specify the feature used.
Until now, using an empty selector in a Send/Fetch request could
accidentally cause messages to be sent to *every single channel on the
network*, which was not very good.

With this patch, any Send/Fetch request MUST now specify at least the
Feature used. The first effect is to effectively forbid empty
selectors, which solves the issue at hand. The second effect is that
this opens the way towards offering channels that implement several
Features at once, e.g. "oven/temperature-c" and "oven-temperature-f",
or "camera/image-jpg" and "camera/image-png".